Move _augment_schema to a method of MerlinDataLoader
